+++ app/javascript/vue/tasks/nomenclature/match/app.vue
@@ -19,7 +19,7 @@
         <input
           type="checkbox"
           v-model="exact"
-        >
+        />
         Exact match
       </label>
     </div>
@@ -36,7 +36,7 @@
                 type="radio"
                 :value="value"
                 v-model="tableView"
-              >
+              />
               {{ value }}
             </label>
           </li>
@@ -49,12 +49,18 @@
     </navbar-component>
 
     <TableUnmatched
-      v-if="(tableView === tableComponent.Unmatched || tableView === tableComponent.Both)"
+      v-if="
+        tableView === tableComponent.Unmatched ||
+        tableView === tableComponent.Both
+      "
       :list="unmatched"
       class="full_width margin-medium-bottom"
     />
     <TableMatched
-      v-if="(tableView === tableComponent.Matches || tableView === tableComponent.Both)"
+      v-if="
+        tableView === tableComponent.Matches ||
+        tableView === tableComponent.Both
+      "
       :valid-names="validTaxonNames"
       :list="matches"
       class="full_width"
@@ -94,41 +100,45 @@ const lines = ref([])
 const tableView = ref(tableComponent.Both)
 const validTaxonNames = ref({})
 
-function GetMatches () {
-  const requests = lines.value.map(name => TaxonName.where({ name, exact: exact.value }))
+function GetMatches() {
+  const requests = lines.value.map((name) =>
+    TaxonName.where({ name, exact: exact.value })
+  )
 
   isLoading.value = true
-  Promise.all(requests).then(async responses => {
-    const taxonNames = [].concat(...responses.map(r => r.body))
-    const uniqueList = getUnique(taxonNames, 'id')
-    const validTaxonNameIDs = uniqueList
-      .filter(taxon => !taxon.cached_is_valid)
-      .map(taxon => taxon.cached_valid_taxon_name_id)
+  Promise.all(requests)
+    .then(async (responses) => {
+      const taxonNames = [].concat(...responses.map((r) => r.body))
+      const uniqueList = getUnique(taxonNames, 'id')
+      const validTaxonNameIDs = uniqueList
+        .filter((taxon) => !taxon.cached_is_valid)
+        .map((taxon) => taxon.cached_valid_taxon_name_id)
 
-    const validNames = validTaxonNameIDs.length
-      ? (await TaxonName.where({ taxon_name_id: validTaxonNameIDs })).body
-      : []
+      const validNames = validTaxonNameIDs.length
+        ? (await TaxonName.where({ taxon_name_id: validTaxonNameIDs })).body
+        : []
 
-    validNames.forEach(taxon => {
-      validTaxonNames.value[taxon.id] = taxon
-    })
-
-    matches.value = uniqueList.map(taxon =>
-      ({
-        taxon,
-        match: lines.value.filter(line => taxon.object_label.toLowerCase().includes(line.toLowerCase())).join(', ')
+      validNames.forEach((taxon) => {
+        validTaxonNames.value[taxon.id] = taxon
       })
-    )
 
-    unmatched.value = lines.value.filter(
-      line => !matches.value.length || !!matches.value.find(({ match }) => !match.includes(line))
-    )
+      matches.value = uniqueList.map((taxon) => ({
+        taxon,
+        match: lines.value.filter((line) =>
+          taxon.object_label.toLowerCase().includes(line.toLowerCase())
+        )
+      }))
 
-    tableView.value = tableComponent.Both
-  })
+      unmatched.value = matches.value.length
+        ? lines.value.filter(
+            (line) => !matches.value.some(({ match }) => match.includes(line))
+          )
+        : lines.value
+
+      tableView.value = tableComponent.Both
+    })
     .finally(() => {
       isLoading.value = false
     })
 }
-
 </script>