+++ app/javascript/vue/tasks/nomenclature/match/components/Table/TableMatched.vue
@@ -9,10 +9,10 @@
     </thead>
     <tbody>
       <tr
-        v-for="({ taxon, match }) in list"
+        v-for="{ taxon, match } in list"
         :key="taxon.id"
       >
-        <td>{{ match }}</td>
+        <td>{{ match.join(', ') }}</td>
         <td>
           <a :href="makeBrowseNomenclatureLink(taxon.id)">
             <span v-html="taxon.object_tag" /> {{ taxon.cached_author_year }}
@@ -30,8 +30,12 @@
             v-if="validNames[taxon.cached_valid_taxon_name_id]"
             :href="makeBrowseNomenclatureLink(taxon.cached_valid_taxon_name_id)"
           >
-            <span v-html="validNames[taxon.cached_valid_taxon_name_id].object_tag" />
-            {{ validNames[taxon.cached_valid_taxon_name_id].cached_author_year }}
+            <span
+              v-html="validNames[taxon.cached_valid_taxon_name_id].object_tag"
+            />
+            {{
+              validNames[taxon.cached_valid_taxon_name_id].cached_author_year
+            }}
           </a>
         </td>
       </tr>
@@ -53,6 +57,6 @@ defineProps({
   }
 })
 
-const makeBrowseNomenclatureLink = (id) => `${RouteNames.BrowseNomenclature}?taxon_name_id=${id}`
-
+const makeBrowseNomenclatureLink = (id) =>
+  `${RouteNames.BrowseNomenclature}?taxon_name_id=${id}`
 </script>